requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# webserver.py: 21
asyncio_redis == 0.14.3

# create_table.py: 1
# rq_worker.py: 1
# webserver.py: 4,13
boto3 == 1.7.45

# webserver.py: 22
elasticsearch == 6.3.0

# extractAudio.py: 1
ffmpeg_python == 0.1.11

# rq_worker.py: 4
ffprobe3 == 0.1.2

# webserver.py: 6
redis == 2.10.6

# webserver.py: 7
rq == 0.11.0

# webserver.py: 15,16
sanic == 0.7.0

# webserver.py: 9
sanic_jinja2 == 0.7.0

# webserver.py: 18
sanic_session == 0.1.3
